backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
requests==2.32.0
python-dotenv==1.0.0
pydantic==2.6.0
google-generativeai==0.13.0
openpyxl==3.2.2
